Remove menu scene
Remove the menu scene entirely. Move this project to only contain the
Phaser tutorial game. This is being done because it was difficult for
me to find a clean way to separate games. Separating them into scenes
isn't great because either every game has only one scene OR I had all
scenes to the game config, which would require changes to the menu
list. Destroying the current game and starting a `new Phaser.Game()`
also proved to be challenging.